third_party/WebKit/LayoutTests/external/wpt/magnetometer/Magnetometer.https-expected.txt

This is a testharness.js-based test.
FAIL Magnetometer: Test that 'onreading' is called and sensor reading is valid assert_equals: Expected reading event, but got error event instead expected "reading" but got "error"
FAIL Magnetometer: sensor reading is correct assert_equals: Expected reading event, but got error event instead expected "reading" but got "error"
FAIL Magnetometer: sensor timestamp is updated when time passes assert_equals: Expected reading event, but got error event instead expected "reading" but got "error"
FAIL Magnetometer: Test that sensor can be successfully created and its states are correct. assert_equals: Expected activate event, but got error event instead expected "activate" but got "error"
FAIL Magnetometer: sensor.start() returns undefined assert_equals: Expected activate event, but got error event instead expected "activate" but got "error"
FAIL Magnetometer: no exception is thrown when calling start() on already started sensor assert_equals: Expected activate event, but got error event instead expected "activate" but got "error"
FAIL Magnetometer: sensor.stop() returns undefined assert_equals: Expected activate event, but got error event instead expected "activate" but got "error"
FAIL Magnetometer: no exception is thrown when calling stop() on already stopped sensor assert_equals: Expected activate event, but got error event instead expected "activate" but got "error"
PASS Magnetometer: throw a 'SecurityError' when constructing sensor object within iframe
FAIL Magnetometer: sensor readings can not be fired on the background tab assert_equals: Expected reading event, but got error event instead expected "reading" but got "error"
Harness: the test ran to completion.
